fix core tests

+++ semtypes/src/main/java/io/ballerina/semtype/subtypedata/IntSubtype.java
@@ -105,8 +105,8 @@ public class IntSubtype implements ProperSubtypeData {
         }
 
         Range r = v.ranges[v.ranges.length - 1];
-        int i = 8;
-        while (i <= 32) {
+        long i = 8L;
+        while (i <= 32L) {
             if (r.max < (1L << i)) {
                 IntSubtype w = createSingleRangeSubtype(0L, i);
                 return w;
